Moved files to be more modular
