--- styles/ColourBoxContainer.module.css
@@ -1,3 +1,0 @@
-/* .ColourBoxContainer {
-  border: 2px solid black;
-} */
